Fix xxsel parsing error.
The xxsel instruction uses part of the standard opc2 field to specify
a additional operand or other values.  A subset of the field is used for
the actual opcode.  The masking and array lookup was getting confused by
bits in the the additional operand field.  The arrays were split so only
the opcodes that should be found for a given mask is in the array.  This
also speeds up the search as you are not searching through values that
cannot match.  The small groups of opcodes for a couple of the masks are
now done in a case statement as that is probably faster then doing an array
look up.

Bugzilla 148000


git-svn-id: svn://svn.valgrind.org/vex/trunk@3284